V0.2.3 final (#46)
* 24 config management (#25)

* initial conifig with conifigurations

* baseline cofigs for configuration management done; all returns are made, but need to adjust and ask for more responses. Pushes are based on folder structure so I check that the service is available to push and submit a descriptiont that doesn't have an audit trail, but i'm tring to discuss with Palo Alto to get that audit trail in place by pulling configurations in candidate and through diff code; still a lot of work with the API needed for better support

* baselined pan.dev calls

* fixed issues with config and updated pushes added notes to readme

* fixed ike issues wiht payload adjusted some configuration issues around multiple child commits that get created, but didn't test yet

* added tag support

* updated reeadme

* updated REMOTE FOLDER location to default to Remote Networks, but now can be passed

* updated a few missing params

* missing folder in verify bwth

* baseline configuration and address address_group object added

* updated setup config and fixed issue with configuration child jobs were looking at parent instead

* Fixed config Management and returns checks of all children and exits out returns value if something was unsuccessful

* updated readme

* Feature/new startup (#26)

* Added autotag create and checks to ensure message body is correct

* changed how to pass auth so that multiple tenants can be used

* cleaned up remote_networks

* Change to startup actions

updated readme updated configuration and config management; fixed error response when error is returned


* clean up

* clean up


* Feature/new startup (#27)

* Added autotag create and checks to ensure message body is correct

* changed how to pass auth so that multiple tenants can be used

* cleaned up remote_networks

* Change to startup actions

updated readme updated configuration and config management; fixed error response when error is returned

* updates to auth

fixed issue with auth and moved to a single location

* clean up

* clean up


* Feature/objects (#28)

* updates to address object creation and udpates to auth around all objects and issues still to be resolved with autotag

* added address get by id

* added address edit functionlty; adjusted the address get by id put_object param

* added error handling in addresses

* Fixed api calls to use auth.verify; updated configs to add additional url endpoint support (#29)

* fixed lenght of tag to be less than or equal to 64 based on docs; still need to fix that each tag is less than or equal to 127 (#30)

* streamlined variables and response when creating a Remote Network (#31)

* added address delete (#32)

* Force pre shared key (#33)

* added address delete

* updated pre-shared-key

* removed debug updated pre-shared-key method

* Final release; cleaned up README.md and removed any remaining DEBUG messages (#34)

* Fix wheel bug (#36)

* updated wheel fix added notes to readme around addresses

* updated version

* Add version to commit (#37)

* added versioning info to commit; updated readme with more examples

* added service connection workspace

* Hotfix bool issue and tunnel monitor (#38)

* Added functionality to get all tags based on total amounts and continue to offset them as they go along

* Added auto tag functionality for creation based off proper payload; had to adjust enumeration of target and add log_type specification in the data

* Fixed issue with bool|str type and set defaults and fixed issue when no tunnel_monitor is wanted

* V0.2.2 final (#39)

* final release of 0.2.2

* updated addresses_edit missing url_type

* added raising an error on address Delete (#43)

* Fixed issue with config not passing correclty when auth isn't passed, updated and fixed address issues (#44)

* cleanup
