+++ prismasase/_version.py
@@ -1,3 +1,3 @@
 """version"""
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
