+++ prismasase/service_setup/ike/ike_crypto.py
@@ -1,9 +1,8 @@
 """IKE Crypto"""
 
-from prismasase.config import Auth
+from prismasase import return_auth
+from prismasase.configs import Auth
 from prismasase.restapi import prisma_request
-from prismasase.utilities import return_auth
-
 
 def ike_crypto_profiles_get(ike_crypto_profile: str, folder: dict, **kwargs) -> str:
     """Checks if IKE Crypto Profile Exists
